+++ cmf-cli/resources/template_feed/iot/taskLibrary/src/%directoryname%/src/index.ts
@@ -1,0 +1,6 @@
+// Export each available task and converter here
+// This file is only used when the library is running in development
+// When the library is packed, this file will be replaced with the
+// full contents and packed
+
+